Refactor promisify logic out of util into new internal module
